remove devel folder
In #5473 we saw a user being confused by devel's continued existance.
No bazel stuff will work any more and there's value in having just one
place where our e2e setup scripts are located, and that's now in `make`.

It was considered that we might keep some of these scripts but
converting them to call make commands, but that might be a lot of work
for minimal reward.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>