make: uncompress pebble outside of the Go Workspace zone of influence
When building Pebble, Go was mistakenly thinking that the go.work is the
Go Workspace in which the Pebble module resides:

  main module (github.com/cert-manager/cert-manager) does not contain package github.com/cert-manager/cert-manager/_bin/downloaded/containers/amd64/pebble/pebble-ba5f81dd80fa870cbc19326f2d5a46f45f0b5ee3/cmd/pebble

At first, I thought that directories prefixed with _ would be ignored
(like what "go build" does), but it doesn't seem to work that way since
the go.work file is looked up recursively "upwards", not downwards.

The only workaround I could think of is to build Pebble outside of the
tree in which go.work resides.

Signed-off-by: Maël Valais <mael@vls.dev>